refactor(controller): inline oci_version_warn helpers into controller.go
Per @brettmthompson's review on #5558: the per-controller oci_version_warn.go
files were thin wrappers around utils.CheckImageVolumeCompatibility. Moved
warnIfImageVolumeUnsupported (and its serverVersioner interface +
OciImageVolumeCompatible condition constant) into the respective controller.go
files as private functions. Tests moved into the controller test surface.

Deleted:
- pkg/controller/v1beta1/inferenceservice/oci_version_warn.go
- pkg/controller/v1beta1/inferenceservice/oci_version_warn_test.go
- pkg/controller/v1alpha2/llmisvc/oci_version_warn.go
- pkg/controller/v1alpha2/llmisvc/oci_version_warn_test.go

The version check logic stays centralized in pkg/utils.CheckImageVolumeCompatibility;
each controller handles its own condition translation inline. Didn't move the
condition translation into pkg/utils itself — that would require utils to import
pkg/apis/serving types, which is an uncommon dependency direction in this codebase.
