+++ pkg/controller/v1beta1/inferenceservice/controller.go
@@ -35,6 +35,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/record"
@@ -769,4 +770,54 @@ func (r *InferenceServiceReconciler) GetFailConditions(isvc *v1beta1.InferenceSe
 		}
 	}
 	return msg
+}
+
+// OciImageVolumeCompatible is an advisory condition surfaced on an InferenceService
+// when native OCI ImageVolume mode is in use and the cluster Kubernetes version may
+// not support it. It never affects the Ready condition (not in conditionSet).
+const OciImageVolumeCompatible apis.ConditionType = "OciImageVolumeCompatible"
+
+// serverVersioner is the subset of discovery.DiscoveryInterface required by
+// warnIfImageVolumeUnsupported. Using a minimal interface enables injection of
+// a lightweight fake in unit tests without implementing all ~30 discovery methods.
+type serverVersioner interface {
+	ServerVersion() (*version.Info, error)
+}
+
+// warnIfImageVolumeUnsupported sets an advisory condition on isvc when the resolved
+// storage mode is "native" and the cluster does not have ImageVolume enabled by default.
+// The compatibility thresholds and version discovery are handled by the shared helper
+// utils.CheckImageVolumeCompatibility; this function translates the result into the
+// ISVC condition format.
+func warnIfImageVolumeUnsupported(ctx context.Context, sv serverVersioner, isvc *v1beta1.InferenceService, resolvedMode string) {
+	if resolvedMode != kservetypes.OciModelModeNative {
+		isvc.Status.ClearCondition(OciImageVolumeCompatible)
+		return
+	}
+
+	result := utils.CheckImageVolumeCompatibility(ctx, sv)
+
+	switch result.Status {
+	case utils.ImageVolumeUnsupported:
+		isvc.Status.SetCondition(OciImageVolumeCompatible, &apis.Condition{
+			Type:   OciImageVolumeCompatible,
+			Status: corev1.ConditionFalse,
+			Reason: "ImageVolumeUnsupported",
+			Message: fmt.Sprintf(
+				"Cluster K8s %s.%s does not support ImageVolume (introduced in 1.31 as alpha). Falling back to modelcar may be required.",
+				result.Major, result.Minor),
+		})
+	case utils.ImageVolumeNeedsGate:
+		isvc.Status.SetCondition(OciImageVolumeCompatible, &apis.Condition{
+			Type:   OciImageVolumeCompatible,
+			Status: corev1.ConditionFalse,
+			Reason: "ImageVolumeAlpha",
+			Message: fmt.Sprintf(
+				"Cluster K8s %s.%s has ImageVolume feature-gated (K8s 1.31–1.34). Ensure --feature-gates=ImageVolume=true is set on kube-apiserver and kubelet.",
+				result.Major, result.Minor),
+		})
+	default:
+		// ImageVolumeOK (≥ 1.35) or ImageVolumeUnknown — clear any previously set warning.
+		isvc.Status.ClearCondition(OciImageVolumeCompatible)
+	}
 }
